+++ src/app/home/home.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { LevelSelectCardsComponent } from '../level-select-cards/level-select-cards.component';
+import { LevelSelectCardsComponent } from './level-select-cards/level-select-cards.component';
 import { SlovnikService } from "../slovnik.service";
 import { RouterLink } from '@angular/router';
