+++ src/app/home/level-select-cards/level-select-cards.component.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../variables';
 
 .level-select-cards {
