+++ src/app/learn/learn.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { LevelSelectCardsComponent } from '../level-select-cards/level-select-cards.component';
+import { LevelSelectCardsComponent } from '../home/level-select-cards/level-select-cards.component';
 
 @Component({
     selector: 'app-learn-page',
